bookshelf/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-filter==2.2.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
freeze==1.0.10
idna==2.8
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
